Move metadata convertion methods to utils
